# backendStub/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API reference of The Spoon project.\n \n VERSIONING: The\
    \ version number works in this way: it's composed of two numbers separated by\
    \ a point. The first number identifies the Sprint, so for example the version\
    \ 2.1 of the API reference is related to the Sprint number 2. The second number\
    \ starts from 0 and has to be increased by one each time the file is modified.\
    \ The person who modifies this document is responsible for increasing the number.\n\
    \ \n GROUPING OF ENDPOINTS: The endpoints are grouped by dividing them into customers'\
    \ operations and owners' operations. All the endpoints that are not strictly related\
    \ to customers nor owners remain inside the default group.\n \n AUTHENTICATION/AUTHORIZATION:\
    \ The json web token returned when the login is performed is supposed to be put\
    \ into x-auth-token header in the endpoints that require authorization"
  version: "6.4"
  title: "The Spoon API reference"
...
paths:
...
  /api/user/customer/review:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Return all the reviews of the customer"
      description: "Return all the reviews of the customer, with their reviewIDs and\
        \ their status (accepted/refused/pending). Returns an empty array if no review\
        \ is found."
      operationId: "getOwnReviews"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MenuReviewWithIDAndStatus"
        "401":
          description: "Access denied"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
  /api/user/customer/review/{reviewID}:
    delete:
      tags:
      - "customer"
      - "Marin"
      summary: "Delete a review"
      description: "Delete the review with the given reviewID. This will also delete\
        \ the reviews of the menu items associated to the review with given reviewID."
      operationId: "apiUserCustomerReviewReviewIDDELETE"
      parameters:
      - name: "reviewID"
        in: "path"
        description: "ID of the review"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Access denied"
        "404":
          description: "Review associated to the customer with given reviewID not\
            \ found"
      security:
      - Bearer: []
      x-security-scopes:
      - "customer"
      x-swagger-router-controller: "Customer"
...
  /api/user/owner/restaurant/review:
    get:
      tags:
      - "owner"
      - "Marin"
      summary: "Return pending reviews"
      description: "Return all the pending reviews of the owner's restaurant. Only\
        \ the image of the receipt is sent (the link to download it from the cloud),\
        \ together with the name of the reviewed menu and the list of the reviewed\
        \ menu items.The pending reviews are sent in an array, which will be empty\
        \ in case there are no pending reviews. With a POST the restaurant owner will\
        \ approve or disapprove the review."
      operationId: "apiUserOwnerRestaurantReviewGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PendingReview"
        "401":
          description: "Access denied"
      security:
      - Bearer: []
      x-security-scopes:
      - "owner"
      x-swagger-router-controller: "Owner"
  /api/user/owner/restaurant/review/{reviewID}:
    post:
      tags:
      - "owner"
      - "Marin"
      summary: "Approve or disapprove pending review"
      description: "Submit the decision of the restaurant owner about the pending\
        \ review with given reviewID (approved or disapproved). In case of a successful\
        \ operation, an array containing all the pending reviews is sent, so that\
        \ the frontend is able to refresh the list (the array sent is like the array\
        \ sent with the GET endpoint)."
      operationId: "apiUserOwnerRestaurantReviewReviewIDPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "reviewID"
        in: "path"
        description: "ID of the pending review"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Submitted status of pending review (approved or disapproved)"
        required: true
        schema:
          $ref: "#/definitions/ApprovalStatus"
      responses:
        "201":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PendingReview"
        "401":
          description: "Access denied"
        "404":
          description: "Review with given reviewID not found"
      security:
      - Bearer: []
      x-security-scopes:
      - "owner"
      x-swagger-router-controller: "Owner"
  /api/user/customer/menu/{menuID}/menuItem/{menuItemID}/review:
    get:
      tags:
      - "customer"
      - "Frikk"
      summary: "Return all the reviews of the menu item"
      description: "Return all the reviews of the menu item with given menuItemID\
        \ contained in the menu with given menuID."
      operationId: "apiUserCustomerMenuMenuIDMenuItemMenuItemIDReviewGET"
      produces:
      - "application/json"
      parameters:
      - name: "menuID"
        in: "path"
        description: "ID of the menu"
        required: true
        type: "integer"
      - name: "menuItemID"
        in: "path"
        description: "ID of the menu item"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ReviewsOfMenuItemWithoutOverall"
        "404":
          description: "Menu or menu item not found"
      x-swagger-router-controller: "Customer"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "x-auth-token"
    in: "header"
definitions:
...
  MenuItemReview:
    type: "object"
    properties:
      menuItemID:
        type: "integer"
      rating:
        type: "number"
      content:
        type: "string"
  MenuReviewWithIDAndStatus:
    type: "object"
    properties:
      menuReviewID:
        type: "integer"
      menuID:
        type: "integer"
      menuName:
        type: "string"
      restaurantName:
        type: "string"
      serviceRating:
        type: "number"
      qualityOverPriceRating:
        type: "number"
      date:
        type: "string"
      receiptImageID:
        type: "integer"
      status:
        type: "string"
      menuItemsReviews:
        type: "array"
        items:
          $ref: "#/definitions/MenuItemReview"
    example:
      menuReviewID: 34342
      menuID: 343
      menuName: "Fish menu"
      restaurantName: "Emilio's Restaurant"
      serviceRating: 4.5999999999999996447286321199499070644378662109375
      qualityOverPriceRating: 4.70000000000000017763568394002504646778106689453125
      date: "2019-12-01"
      receiptImageID: 67
      status: "accepted"
      menuItemsReviews:
      - menuItemID: 34
        rating: 4.0999999999999996447286321199499070644378662109375
        content: "That was delicious!"
      - menuItemID: 58
        rating: 0.5
        content: "It was an insult, I will never eat that trash again in my whole\
          \ life"
...
  PendingReview:
    type: "object"
    properties:
      reviewID:
        type: "integer"
      menuName:
        type: "string"
      menuItemNames:
        type: "array"
        items:
          $ref: "#/definitions/PendingReview_menuItemNames"
    example:
      reviewID: 988
      menuName: "Sea menu"
      menuItemNames:
      - menuItemName: "Spaghetti allo scoglio"
      - menuItemName: "Sashimi"
  ApprovalStatus:
    type: "object"
    properties:
      isApproved:
        type: "boolean"
    example:
      isApproved: true
...
  ReviewsOfMenuItemWithoutOverall:
    type: "object"
    properties:
      username:
        type: "string"
      rating:
        type: "number"
      content:
        type: "string"
      date:
        type: "string"
    example:
      username: "Janine"
      rating: 5
      content: "Best pizza I have tasted in ages!"
      date: "2019-12-01"
...
  PendingReview_menuItemNames:
    type: "object"
    properties:
      menuItemName:
        type: "string"